Improved the error and loading states

+++ todo-app/backend/server.js
@@ -15,11 +15,11 @@ app.use(morgan("dev"));
 // GET
 app.get('/api', async (req, res) => {
   try {
-    const data = await pool.query("SELECT * FROM todos ORDER BY id");
+    const data = await pool.query("SELECT * FROM todos WHERE user_id = 1 ORDER BY id");
     res.json(data.rows);
   } catch (err) {
     console.error(err);
-    res.status(500).json({ error: "Interner Serverfehler" });
+    res.status(500).json({ error: "Todos konnten nicht geladen werden. Versuche es später erneut" });
   }
 })
 
@@ -31,18 +31,24 @@ app.post('/api', async (req, res) => {
 
     const entry = req.body;
 
-    // Prüfen, ob der Body oder ein einzelnes Feld leer sind
-    if (!entry || entry === {}) {
-      return res.status(400).json({ error: "Der Todo-Eintrag darf nicht leer sein" });
-    }
-    if (!entry.name || entry.name.trim() === "" || !entry.deadline || entry.deadline.trim() === "") {
-      return res.status(400).json({ error: "Alle Felder des neuen Todo-Eintrags müssen ausgefüllt sein" });
+    // Prüfen, ob der Body oder alle Felder leer sind
+    if (!entry || entry === {} || ((!entry.name || entry.name.trim() === "") && (!entry.deadline || entry.deadline.trim() === ""))) {
+      return res.status(400).json({ name: "Bitte gib einen Namen für den Todo-Eintrag an", deadline: "Bitte gib eine Deadline für den Todo-Eintrag an"  });
     }
 
+    // Prüfen, ob der neue Name gültig ist
+    if (!entry.name || entry.name.trim() === "") {
+      return res.status(400).json({ name: "Bitte gib einen Namen für den Todo-Eintrag an" });
+    }
     // Überprüfen, ob ein solcher Eintrag bereits existiert
-    const query = await pool.query("SELECT * FROM todos WHERE name = $1", [entry.name]);
+    const query = await pool.query("SELECT * FROM todos WHERE name = $1 AND user_id = 1", [entry.name]);
     if (query.rowCount !== 0) {
-      return res.status(409).json({ error: "Es gibt bereits ein Todo-Eintrag mit diesem Namen" });
+      return res.status(409).json({ name: "Es gibt bereits ein Todo-Eintrag mit diesem Namen" });
+    }
+
+    // Prüfen, ob die neue Deadline gültig ist
+    if (!entry.deadline || entry.deadline.trim() === "") {
+      return res.status(400).json({ deadline: "Bitte gib eine Deadline für den Todo-Eintrag an" });
     }
 
     // Neuen Eintrag zur Datenbank hinzufügen
@@ -50,13 +56,13 @@ app.post('/api', async (req, res) => {
     console.log("Neuer Eintrag hinzugefügt: " + entry.name);
 
     // Erfolgsmeldung
-    const id = await pool.query("SELECT id FROM todos WHERE name = $1", [entry.name]);
+    const id = await pool.query("SELECT id FROM todos WHERE name = $1 AND user_id = 1", [entry.name]);
     res.status(201).json({ id: id, name: entry.name, deadline: entry.deadline });
   } catch (err) {
 
     // Fehlermeldung
     console.error(err);
-    return res.status(500).json({ error: "Todo konnte nicht erstellt werden. Versuche es später erneut" });
+    return res.status(500).json({ general: "Todo konnte nicht erstellt werden. Versuche es später erneut" });
   }
 });
 
@@ -69,13 +75,13 @@ app.delete('/api/:id', async (req, res) => {
     const id = parseInt(req.params.id);
 
     // Prüfen, ob das Element existiert, das gelöscht werden soll
-    const query = await pool.query("SELECT * FROM todos WHERE id = $1", [id]);
+    const query = await pool.query("SELECT * FROM todos WHERE id = $1 AND user_id = 1", [id]);
     if (!query.rowCount === 0) {
       return res.status(404).json({ error: "Der zu löschende Todo-Eintrag wurde nicht gefunden" });
     }
 
     // Das Element aus der Datenbank löschen
-    await pool.query("DELETE FROM todos WHERE id = $1", [id])
+    await pool.query("DELETE FROM todos WHERE id = $1 AND user_id = 1", [id])
 
     // Erfolgsmeldung
     res.status(201).json(id);
@@ -97,32 +103,34 @@ app.patch('/api/:id', async (req, res) => {
   try {
 
     // Prüfen, ob das Element existiert, das verändert werden soll
-    const query = await pool.query("SELECT * FROM todos WHERE id = $1", [id]);
+    const query = await pool.query("SELECT * FROM todos WHERE id = $1 AND user_id = 1", [id]);
     if (!query.rowCount === 0) {
-      return res.status(404).json({ error: "Der zu bearbeitende Todo-Eintrag wurde nicht gefunden" });
+      return res.status(404).json({ general: "Der zu bearbeitende Todo-Eintrag wurde nicht gefunden" });
     }
 
     // Prüfen, ob überhaupt Änderungen angegeben wurden
     if (!newName && !newDeadline) {
-      return res.status(400).json({ error: "Keine Änderung angegeben" });
+      return res.status(400).json({ general: "Keine gültige Änderung angegeben" });
     }
 
     // Prüfen, ob der angegebene neue Name gültig ist
-    if (newName !== undefined) {
+    const oldName = query.rows[0].name;
+    if (newName !== oldName) {
       if (typeof newName !== "string" || newName.trim() === "") {
-        return res.status(400).json({ error: "Der mitgeschickte neue Name ist ungültig" });
+        return res.status(400).json({ name: "Der neue Name ist ungültig" });
       }
 
-      await pool.query("Update todos SET name = $1 WHERE id = $2", [newName, id]);
+      await pool.query("Update todos SET name = $1 WHERE id = $2 AND user_id = 1", [newName, id]);
     }
 
     // Prüfen, ob die angegebene neue Deadline gültig ist
+    const oldDeadline = query.rows[0].deadline;
     if (newDeadline !== undefined) {
       if (typeof newDeadline !== "string" || newDeadline.trim() === "") {
-        return res.status(400).json({ error: "Die mitgeschickte neue Deadline ist ungültig" });
+        return res.status(400).json({ deadline: "Die neue Deadline ist ungültig" });
       }
 
-      await pool.query("Update todos SET deadline = $1 WHERE id = $2", [newDeadline, id]);
+      await pool.query("Update todos SET deadline = $1 WHERE id = $2 AND user_id = 1", [newDeadline, id]);
     }
 
     // Erfolgsmeldung
@@ -131,7 +139,7 @@ app.patch('/api/:id', async (req, res) => {
 
     // Fehlermeldung
     console.error(err);
-    return res.status(500).json({ error: "Todo konnte nicht erstellt werden. Versuche es später erneut" });
+    return res.status(500).json({ general: "Todo konnte nicht erstellt werden. Versuche es später erneut" });
   }
 });
 
@@ -143,25 +151,29 @@ app.post('/api/register', async (req, res) => {
     const username = req.body.username.trim();
     const password = req.body.password;
 
-    // Prüfen, ob der Body oder ein einzelnes Feld leer sind
-    if (!req.body || req.body === {} || !username || username.trim() === "" || !password || password.trim() === "") {
-      return res.status(400).json({ error: "Alle Felder müssen ausgefüllt sein" });
+    // Prüfen, ob der Body oder alle Felder leer sind
+    if (!req.body || req.body === {} || (!username || (username.trim() === "") && (!password || password.trim() === ""))) {
+      return res.status(400).json({ username: "Bitte gib einen Username an", password: "Bitte gib ein Passwort an" });
     }
 
-    // Prüfen, ob der Username bereits existiert
+    // Prüfen, ob der Username gültig ist
+    if (!username || username.trim() === "") {
+      return res.status(400).json({ username: "Bitte gib einen gültigen Username an" });
+    }
     const query = await pool.query("SELECT * FROM users WHERE username = $1", [username]);
     if (query.rowCount !== 0) {
-      return res.status(409).json({ error: "Dieser Name ist bereits vergeben" });
+      return res.status(409).json({ username: "Dieser Username ist bereits vergeben" });
     }
-
-    // Prüfen, ob der Username die richtige Länge besitzt
     if (username.length < 3 || username.length > 30) {
-      return res.status(400).json({ error: "Der Username muss zwischen 3 und 30 Zeichen lang sein" });
+      return res.status(400).json({ username: "Der Username muss zwischen 3 und 30 Zeichen lang sein" });
     }
 
-    // Prüfen, ob das Passwort die richtige Länge besitzt
+    // Prüfen, ob das Passwort gültig ist
+    if (!password || password.trim() === "") {
+      return res.status(400).json({ password: "Bitte gib ein gültiges Passwort an" });
+    }
     if (password.length < 8 || password.length > 100) {
-      return res.status(400).json({ error: "Das Passwort muss zwischen 8 und 100 Zeichen lang sein" });
+      return res.status(400).json({ password: "Das Passwort muss zwischen 8 und 100 Zeichen lang sein" });
     }
 
     // Passwort hashen
@@ -171,12 +183,12 @@ app.post('/api/register', async (req, res) => {
     await pool.query("INSERT INTO users(username, password_hash) VALUES ($1, $2)", [username, passwordHash])
 
     // Erfolgsmeldung
-    res.status(201).json({username, passwordHash});
+    res.status(201).json({ username, passwordHash });
   } catch (err) {
 
     // Fehlermeldung
     console.error(err);
-    return res.status(500).json({ error: "Todo konnte nicht gelöscht werden. Versuche es später erneut" });
+    return res.status(500).json({ general: "Der Account konnte nicht erstellt werden. Probiere es später erneut" });
   }
 });
 
@@ -188,27 +200,34 @@ app.post('/api/login', async (req, res) => {
     const username = req.body.username.trim();
     const password = req.body.password;
 
-    // Prüfen, ob der Body oder ein einzelnes Feld leer sind
-    if (!req.body || req.body === {} || !username || username.trim() === "" || !password || password.trim() === "") {
-      return res.status(400).json({ error: "Alle Felder müssen ausgefüllt sein" });
+    // Prüfen, ob der Body oder alle Felder leer sind
+    if (!req.body || req.body === {} || (!username || username === "") && (!password || password.trim() === "")) {
+      return res.status(400).json({ username: "Bitte gib einen gültigen Username an", password: "Bitte gib ein gültiges Passwort an" });
     }
 
-    // User abrufen
+    // Prüfen, ob der Username gültig ist
+    if (!username || username.trim() === "") {
+      return res.status(400).json({ username: "Bitte gib einen gültigen Username an" });
+    }
     const user = await pool.query("SELECT * FROM users WHERE username = $1", [username]);
     if (user.rowCount === 0) {
-      return res.status(404).json({ error: "Der eingegebene Username existiert nicht" });
+      return res.status(404).json({ username: "Der eingegebene Username existiert nicht" });
     }
 
-    // Passwörter vergleichen
-    if (!bcrypt.compare(password, user.rows.password_hash)) {
-      return res.status(404).json({ error: "Username und Passwort stimmen nicht überein" });
+    // Prüfen, ob das Passwort gültig ist
+    if (!password || password.trim() === "") {
+      return res.status(400).json({ password: "Bitte gib ein gültiges Passwort an" });
+    }
+    const passwordMatch = await bcrypt.compare(password, user.rows[0].password_hash);
+    if (!passwordMatch) {
+      return res.status(401).json({ general: "Username und Passwort stimmen nicht überein" });
     }
 
     // JWT generieren
     const secret = process.env.JWT_SECRET;
     const token = jwt.sign(
       {
-        userId: user.rows.id
+        userId: user.rows[0].id
       },
       process.env.JWT_SECRET,
       {
@@ -222,7 +241,7 @@ app.post('/api/login', async (req, res) => {
 
     // Fehlermeldung
     console.error(err);
-    return res.status(500).json({ error: "Todo konnte nicht gelöscht werden. Versuche es später erneut" });
+    return res.status(500).json({ general: "Anmeldung fehlgeschlagen. Versuche es später erneut" });
   }
 });
